# Conflicts:
#	src/main/java/com/example/mofmof/controller/TodolistController.java
#	src/main/java/com/example/mofmof/repository/TaskRepository.java
#	src/main/java/com/example/mofmof/repository/entity/Tasks.java

+++ src/main/java/com/example/mofmof/repository/entity/Tasks.java
@@ -21,7 +21,7 @@ public class Tasks {
     private String content;
 
     @Column
-    private int status;
+    private Short status;
 
     @Column(name = "limit_date")
     private Date limitDate;
